Major re-engineering of the stats gathering
Docker's stats API can take up to one second to return the first batch
of stats data, so the way this was implemented breaks down when the
number of running containers becomes close to, or greater than
CollectD's polling interval.

We know implement a continuously-consuming thread of stats for each
container, and get the most recently read stats data from that thread
directly when CollectD wants data.

Signed-off-by: Maxime Petazzoni <max@signalfx.com>