Avoid infinite loops when failing to communicate with Docker for stats
If for some reason the Docker daemon correctly answers the /containers
endpoint, but fails to answer the /stats endpoint, we would create a
stats gathering thread that could go into an infinite tight loop logging
errors.

Change the error handling behavior in the stats gathering thread to wait
one second on failure and abort after three consecutive failures. If the
container really is still there, the stats gathering thread will be
recreated in a subsequent read_callback() call.

Signed-off-by: Maxime Petazzoni <max@signalfx.com>